Create README.md
YTDSALES (Year-to-Date Sales): This KPI tracks the total sales revenue generated by your eCommerce business from the beginning of the current year up to the present. It provides insight into your revenue performance over a specific period.

YTDPROFIT (Year-to-Date Profit): YTDPROFIT measures the total profit your eCommerce business has earned from the start of the current year to the current date. It indicates how well your company is managing costs and generating profits during that time frame.

YTDQTY (Year-to-Date Quantity): YTDQTY represents the total quantity or volume of products or services sold by your eCommerce business from the beginning of the current year to the present. It helps assess sales volume trends.

YTDPROFIT MARGIN: YTDPROFIT MARGIN is a financial KPI that calculates the percentage of profit your business makes on each sale during the current year. It's a crucial metric for understanding the profitability of your products or services.

YOYSALES (Year-over-Year Sales): YOYSALES compares the total sales revenue from the current year to the same period in the previous year. This KPI helps identify growth or decline in sales performance over time.

YOYPROFIT (Year-over-Year Profit): YOYPROFIT measures the difference in profit between the current year and the previous year for the same period. It assesses whether your business is becoming more or less profitable.

YOYQTY (Year-over-Year Quantity): YOYQTY compares the quantity of products or services sold in the current year to the same period in the previous year. It helps gauge changes in sales volume.

YOYPROFIT MARGIN (Year-over-Year Profit Margin): This KPI calculates the percentage change in profit margin between the current year and the previous year for the same period. It provides insight into changes in profitability over time.